+++ app/models/demux/connection.rb
@@ -24,10 +24,14 @@ module Demux
 
     # Return an entry url for this specific connection
     #
-    # @return [String] the entry url with account_id in signed token
+    # @param data [Hash] extra data to pass along when building the entry_url.
+    #   Whatever is passed in this hash will be included in addition to the
+    #   default `account_id` key.
+    #
+    # @return [String] the entry url with account ID and other data in token
 
-    def entry_url
-      app.signed_entry_url(data: { account_id: account_id })
+    def entry_url(data: {})
+      app.signed_entry_url(data: data.merge(account_id: account_id))
     end
   end
 end